added experiments with transcriptions matching the size of japanese data
+ fixed missing run.sh
+ merged all experiments in run.sh

+++ platalea/experiments/config.py
@@ -54,7 +54,7 @@ args.add_argument(
 args.add_argument(
     '--seed', default=123, type=int, help='seed for sources of randomness')
 parser.add_argument(
-    '--downsampling-factor', default=None, type=int,
+    '--downsampling-factor', default=None, type=float,
     dest='downsampling_factor',
     help='factor by which the dataset should be downsampled')
