+++ platalea/experiments/flickr8k/mtl_asr.py
@@ -11,7 +11,7 @@ from platalea.experiments.config import args
 
 # Parsing arguments
 args.add_argument(
-    '--downsampling_factor_text', default=None, type=int,
+    '--downsampling_factor_text', default=None, type=float,
     help='factor by which the amount of available transcriptions should be \
     downsampled (affecting ASR only)')
 args.enable_help()
